+++ feature/notification/src/main/java/com/yangbong/damedame/notification/work_manager/WorkManager.kt
@@ -1,0 +1,4 @@
+package com.yangbong.damedame.notification.work_manager
+
+class WorkManager {
+}
